requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.7
pydantic==2.10.3
jinja2==3.1.2
markdown2==2.4.10
aiofiles==24.1.0
python-magic==0.4.27
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 